rabbit mq

+++ services/auth-service/src/services/logout.service.ts
@@ -5,26 +5,24 @@ import { redis } from "../lib/redis.client";
 import { publishEvent } from "../events/publisher";
 import { ROUTING_KEY } from "../events/routingKey";
 import jwt from "jsonwebtoken";
+import { verifyAccessToken } from "../utils/generateToken";
+import { AppError } from "../error/AppError";
 
-export const logoutService = async (
-  accessToken: string,
-  refreshToken: string
-) => {
+export const logoutService = async (accessToken: string, refreshToken: string) => {
   // decode access token (no verify needed here)
-  const decoded: any = jwt.decode(accessToken);
+  const payload  = verifyAccessToken(accessToken);
+  if(!payload){
+    throw new AppError("Invalid access token", 401);
+  }
 
-  const exp = decoded?.exp;
-
-  // 🧠 blacklist access token in redis
-  const ttl = exp - Math.floor(Date.now() / 1000);
+  // calculating remainig time of access token and store in blacklist in redis 
+  const ttl = payload.exp! -Math.floor(Date.now() / 1000);
 
   if (ttl > 0) {
     await redis.set(`bl:${accessToken}`, "true", "EX", ttl);
   }
-
-  // delete refresh session
+  //  delete refresh session
   const hash = await hashToken(refreshToken);
-
   await prisma.session.deleteMany({
     where: {
       tokenHash: hash,
@@ -32,9 +30,9 @@ export const logoutService = async (
   });
 
   // 📡 publish logout event
-  await publishEvent(ROUTING_KEY.AUTH_LOGOUT, {
-    userId: decoded.userId,
-  });
+  // await publishEvent(ROUTING_KEY.AUTH_LOGOUT, {
+  //   userId: payload.userId,
+  // });
 
   return { message: "Logged out successfully" };
 };
